requirements.txt
# Dependencies for master_electronics project
google-cloud-bigquery>=3.0.1
pandas-gbq>=0.17.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
pandas>=1.3.0
asyncio>=3.4.3
asyncpg>=0.25.0
imputemodule @ git+ssh://git@github.com/EETech-Group/impute-module.git@main
serpmodule  @ git+ssh://git@github.com/EETech-Group/serp-module.git@main
extractmodule @ git+ssh://git@github.com/EETech-Group/extract-module.git@main
python-dotenv>=0.21.0
fastapi>=0.95.0
uvicorn[standard]>=0.20.0


# Note: If you encounter issues with SSH access, ensure your SSH key is correctly configured for GitHub
# and that there are no network restrictions preventing access to github.com.
# Alternatively, install locally if the module is available.
